mn10300: fix trap.h errors with gcc-15

+++ libgloss/mn10300/open.c
@@ -8,5 +8,5 @@ int
 _open (const char *path,
 	int flags)
 {
-  return TRAP0 (SYS_open, path, flags, 0);
+  return TRAP0 (SYS_open, (unsigned long) path, flags, 0);
 }
